profile page displays all saved pages from DB
